% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costoms.R
\name{expimp_data}
\alias{expimp_data}
\title{expimp_data() Function}
\usage{
expimp_data(ServiceKey, startpoint, endpoint, hscode, country)
}
\arguments{
\item{a}{,b : two numbers to be operated}
}
\description{
This function adds a and b
}
\examples{
expimp_data()
}
\keyword{Add}
