% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costoms.R
\name{getNitemtradeList}
\alias{getNitemtradeList}
\title{getNitemtradeList}
\usage{
getNitemtradeList(ServiceKey, startpoint, endpoint, hscode, country)
}
\arguments{
\item{ServiceKey}{your service key.}

\item{startpoint}{year,month. character}

\item{endpoint}{year,month. character}

\item{hscode}{4-digit hscode. character}

\item{country}{country code.}
}
\description{
\code{getNitemtradeList} returns Import/Export Performance by Item and Country.
}
\details{
This function return Import/Export Performance by Item and Country.
}
\examples{
getNitemtradeList('yourservicekey', '202001', '202012', '3304', 'US')
}
