% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costoms.R
\name{one_getNitemtradeList}
\alias{one_getNitemtradeList}
\title{request one api}
\usage{
one_getNitemtradeList(ServiceKey, startpoint, endpoint, hscode, country)
}
\arguments{
\item{ServiceKey}{your service key.}

\item{startpoint}{year,month. character}

\item{endpoint}{year,month. character}

\item{hscode}{4-digit hscode. character}

\item{country}{country code.}
}
\description{
request one api
}
